+++ libs/langgraph/src/web.ts
@@ -31,7 +31,8 @@ export {
 } from "./channels/index.js";
 export { type AnnotationRoot as _INTERNAL_ANNOTATION_ROOT } from "./graph/index.js";
 export { type RetryPolicy } from "./pregel/utils/index.js";
-export { Send } from "./constants.js";
+export { Send, Command, type Interrupt } from "./constants.js";
+export { interrupt } from "./interrupt.js";
 
 export {
   MemorySaver,
